+++ RELEASE_NOTES.md
@@ -3,7 +3,7 @@
 ## 1.14
 
 **Fixed**
-- Binary and decimal units are respected when specifying memory. For example, GB is 10<sup>9</sup> bytes, and GiB is 2^30 bytes.
+- Binary and decimal units are respected when specifying memory. For example, GB is 10<sup>9</sup> bytes, and GiB is 2<sup>30</sup> bytes.
 
 **Added**
 - Initial support for the development WDL version, 1.1 (or 2.0). This does not include the directory type.